Refactor: Split BusinessDashboard into components
Refactor the BusinessDashboard page into smaller, reusable components to improve code organization and readability.
